Gatekeeper method working now as an security layer
