+++ lib/bl-security-addons.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'bl-security-addons/version'
+require 'bl-security-addons/gatekeeper'
+
+module BlSecurityAddons
+  class Error < StandardError; end
+  # Your code goes here...
+end
